Move orbit to domain.orbit
