+++ gravitate/data_access/__init__.py
@@ -4,5 +4,4 @@
 
 from .event_schedule_dao import EventScheduleGenericDao
 # from .event_dao import EventDao
-from .orbit_dao import OrbitDao
 from .user_dao import UserDao
