+++ gravitate/domain/event_schedule/actions.py
@@ -3,7 +3,7 @@ from typing import Type
 from gravitate.domain.event_schedule.builders import EventScheduleBuilder
 from gravitate.domain.rides import RideRequest, AirportRideRequest, SocialEventRideRequest
 from gravitate.domain.location.models import Location, AirportLocation, SocialEventLocation
-from gravitate.models import Orbit
+from gravitate.domain.orbit import Orbit
 
 
 def create_event_schedule(ride_request: Type[RideRequest], location: Type[Location], event=None):
