+++ gravitate/models/__init__.py
@@ -1,5 +1,4 @@
 # Import Model Classes
 from .target import Target, ToEventTarget, FromEventTarget
-from .orbit import Orbit
 # from .event import Event, AirportEvent, SocialEvent
 from .user import User
