+++ test/test_domain/test_group/test_orbit_group.py
@@ -6,7 +6,7 @@ import gravitate.domain.group.utils as grouping_utils
 import test.store.model
 from gravitate.domain.rides.dao import RideRequestGenericDao
 from gravitate.domain.group.utils import _add_to_orbit
-from gravitate.models.orbit import Orbit
+from gravitate.domain.orbit.models import Orbit
 from gravitate.domain.rides import RideRequest
 from gravitate import context
 from test import scripts
